refactor: extract torrent file service
to generate random torrent files. It was moved from the handler.
