+++ src/services/mod.rs
@@ -6,4 +6,5 @@ pub mod proxy;
 pub mod settings;
 pub mod tag;
 pub mod torrent;
+pub mod torrent_file;
 pub mod user;
